+++ database/factories/ResourceFactory.php
@@ -1,7 +1,7 @@
 <?php
 
 use Faker\Generator as Faker;
-use GuiFavere\LaravelSimpleScopes\Tests\Resource;
+use GuiFavere\LaravelSimpleScopes\Tests\Models\Resource;
 
 $factory->define(Resource::class, fn (Faker $faker) => [
     'name' => $faker->words(3, true),
