+++ tests/SimpleScopesTest.php
@@ -2,7 +2,10 @@
 
 namespace GuiFavere\LaravelSimpleScopes\Tests;
 
-function fakeRecords(): array {
+use GuiFavere\LaravelSimpleScopes\Tests\Models\Resource;
+
+function fakeRecords(): array
+{
     return [
         ['created_at' => '2023-05-09 09:20:05', 'updated_at' => '2023-05-09 09:20:05'],
         ['created_at' => '2023-05-10 08:20:05', 'updated_at' => '2023-05-10 08:20:05'],
@@ -26,24 +29,24 @@ class SimpleScopesTest extends TestCase
     /** @test */
     public function should_retrieve_all_records_from_a_given_date(): void
     {
-        $this->assertEquals(4, Resource::from('2023-05-10')->get()->count());
+        $this->assertEquals(4, Resource::from('2023-05-10')->count());
     }
 
     /** @test */
     public function should_retrieve_all_records_to_a_given_date(): void
     {
-        $this->assertEquals(1, Resource::to('2023-05-09')->get()->count());
+        $this->assertEquals(1, Resource::to('2023-05-09')->count());
     }
 
     /** @test */
     public function should_retrieve_all_records_modified_from_a_given_date(): void
     {
-        $this->assertEquals(2, Resource::modifiedFrom('2023-05-11')->get()->count());
+        $this->assertEquals(2, Resource::modifiedFrom('2023-05-11')->count());
     }
 
     /** @test */
     public function should_retrieve_all_records_modified_to_a_given_date(): void
     {
-        $this->assertEquals(3, Resource::modifiedTo('2023-05-10')->get()->count());
+        $this->assertEquals(3, Resource::modifiedTo('2023-05-10')->count());
     }
 }
